# .migration-backup/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FinanceDOM API — Subventions non remboursables Outre-Mer
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dossiers:
    get:
      operationId: listDossiers
      tags: [dossiers]
      responses:
        "200":
          description: List of dossiers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dossier"
    post:
      operationId: createDossier
      tags: [dossiers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDossierBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dossier"
  /dossiers/{id}:
    get:
      operationId: getDossier
      tags: [dossiers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dossier detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DossierDetail"
    put:
      operationId: updateDossier
      tags: [dossiers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDossierBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dossier"
  /dossiers/{id}/submit:
    post:
      operationId: submitDossier
      tags: [dossiers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dossier"

  # ── Documents ──────────────────────────────────────
  /dossiers/{id}/documents:
    get:
      operationId: listDocuments
      tags: [documents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: uploadDocument
      tags: [documents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadDocumentBody"
      responses:
        "201":
          description: Uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /dossiers/{dossierId}/documents/{docId}:
    delete:
      operationId: deleteDocument
      tags: [documents]
      parameters:
        - name: dossierId
          in: path
          required: true
          schema:
            type: integer
        - name: docId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Messages ──────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ── Auth ──
...
    Dossier:
      type: object
      properties:
        id:
          type: integer
        reference:
          type: string
        titre:
          type: string
        territoire:
          type: string
        dispositif:
          type: string
        secteur:
          type: string
        statut:
          type: string
        montantDemande:
          type: number
        montantApport:
          type: number
        progressionEtape:
          type: integer
        totalEtapes:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, reference, titre, territoire, dispositif, secteur, statut, montantDemande, progressionEtape, totalEtapes, createdAt, updatedAt]
    DossierDetail:
      allOf:
        - $ref: "#/components/schemas/Dossier"
        - type: object
          properties:
            description:
              type: string
            justificationBudget:
              type: string
            dateDebut:
              type: string
            dureeProjet:
              type: string
            documents:
              type: array
              items:
                $ref: "#/components/schemas/Document"
            messages:
              type: array
              items:
                $ref: "#/components/schemas/Message"
            frais:
              type: array
              items:
                $ref: "#/components/schemas/Frais"
            timeline:
              type: array
              items:
                $ref: "#/components/schemas/TimelineEvent"
    CreateDossierBody:
      type: object
      properties:
        titre:
          type: string
        territoire:
          type: string
        dispositif:
          type: string
        secteur:
          type: string
        description:
          type: string
        montantDemande:
          type: number
        montantApport:
          type: number
        justificationBudget:
          type: string
        dateDebut:
          type: string
        dureeProjet:
          type: string
      required: [titre, territoire, dispositif, secteur, montantDemande]
    UpdateDossierBody:
      type: object
      properties:
        titre:
          type: string
        description:
          type: string
        montantDemande:
          type: number
        montantApport:
          type: number
        justificationBudget:
          type: string
        dateDebut:
          type: string
        dureeProjet:
          type: string

    # ── Documents ──
    Document:
      type: object
      properties:
        id:
          type: integer
        dossierId:
          type: integer
        type:
          type: string
        nom:
          type: string
        filename:
          type: string
        statut:
          type: string
          enum: [en_attente, valide, rejete, manquant]
        obligatoire:
          type: boolean
        uploadedAt:
          type: string
      required: [id, dossierId, type, nom, statut, obligatoire]
    UploadDocumentBody:
      type: object
      properties:
        type:
          type: string
        nom:
          type: string
        filename:
          type: string
      required: [type, nom, filename]

    # ── Messages ──
    Message:
      type: object
      properties:
        id:
          type: integer
        dossierId:
          type: integer
        expediteur:
          type: string
        expediteurRole:
          type: string
          enum: [user, admin, system]
        contenu:
          type: string
        lu:
          type: boolean
        createdAt:
          type: string
      required: [id, dossierId, expediteur, expediteurRole, contenu, lu, createdAt]
...
    Frais:
      type: object
      properties:
        id:
          type: integer
        dossierId:
          type: integer
        reference:
          type: string
        montantHT:
          type: number
        montantTVA:
          type: number
        montantTTC:
          type: number
        statut:
          type: string
          enum: [en_attente, paye, annule]
        echeance:
          type: string
        paidAt:
          type: string
        createdAt:
          type: string
      required: [id, dossierId, reference, montantHT, montantTVA, montantTTC, statut, echeance, createdAt]
...
    TimelineEvent:
      type: object
      properties:
        etape:
          type: integer
        titre:
          type: string
        description:
          type: string
        statut:
          type: string
          enum: [fait, en_cours, en_attente]
        date:
          type: string
      required: [etape, titre, description, statut]

    # ── Admin ──
...